+++ crates/oxc_sourcemap/CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) until v1.0.0.
 
+## [0.23.0] - 2024-08-01
+
+- 27fd062 sourcemap: [**BREAKING**] Avoid passing `Result`s (#4541) (overlookmotel)
+
+### Performance
+
+- d00014e sourcemap: Elide bounds checks in VLQ encoding (#4583) (overlookmotel)
+- 1fd9dd0 sourcemap: Use simd to escape JSON string (#4487) (Brooooooklyn)
+
+### Refactor
+
+- 7c42ffc sourcemap: Align Base64 chars lookup table to cache line (#4535) (overlookmotel)
+
 ## [0.22.1] - 2024-07-27
 
 ### Bug Fixes